% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_template}
\alias{show_template}
\title{Show template}
\usage{
show_template(
  filename,
  lang = "markdown",
  details = FALSE,
  yaml_only = FALSE,
  ...
)
}
\arguments{
\item{filename}{Filename}

\item{lang}{Language (for instance markdown)}

\item{details}{Whether to show as details}

\item{yaml_only}{}

\item{...}{passed to \code{details::details()}}
}
\description{
Show template
}
